+++ SqlCompiler/src/sources.txt
@@ -1,0 +1,5 @@
+./Utils/Utils.java
+./SqlCompiler.java
+./Tokenizer/TokenType.java
+./Tokenizer/Tokenizer.java
+./Tokenizer/Token.java
